Kernel: Debug WriteHex now outputs the high bytes

+++ Kernel/src/Debug.cpp
@@ -18,7 +18,8 @@ void Debug::WriteHex(UInt32 value) const {
 	                        '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'};
 	this->Write("0x");
 	for (UInt8 i = (sizeof(UInt32) * 8 - 4); i > 0; i -= 4) {
-		UInt8 halfByte = (static_cast<uint32_t>(value) >> i) & 0xF;
+		UInt8 halfByte =
+		    (static_cast<uint32_t>(value) >> static_cast<uint8_t>(i)) & 0xF;
 		this->Write(lookup[static_cast<uint8_t>(halfByte)]);
 	}
 	this->Write(lookup[static_cast<uint32_t>(value) & 0xF]);
@@ -29,7 +30,8 @@ void Debug::WriteHex(UIntPtr value) const {
 	                        '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'};
 	this->Write("0x");
 	for (UInt8 i = (sizeof(UIntPtr) * 8 - 4); i > 0; i -= 4) {
-		UInt8 halfByte = (static_cast<uintptr_t>(value) >> i) & 0xF;
+		UInt8 halfByte =
+		    (static_cast<uintptr_t>(value) >> static_cast<uint8_t>(i)) & 0xF;
 		this->Write(lookup[static_cast<uint8_t>(halfByte)]);
 	}
 	this->Write(lookup[static_cast<uintptr_t>(value) & 0xF]);
@@ -40,7 +42,8 @@ void Debug::WriteHex(USize value) const {
 	                        '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'};
 	this->Write("0x");
 	for (UInt8 i = (sizeof(USize) * 8 - 4); i > 0; i -= 4) {
-		UInt8 halfByte = (static_cast<size_t>(value) >> i) & 0xF;
+		UInt8 halfByte =
+		    (static_cast<size_t>(value) >> static_cast<uint8_t>(i)) & 0xF;
 		this->Write(lookup[static_cast<uint8_t>(halfByte)]);
 	}
 	this->Write(lookup[static_cast<size_t>(value) & 0xF]);
